+++ addons/eventsheet/editor/dock/multi_view_manager.gd
@@ -80,6 +80,7 @@ func _connect_view_signals(view: EventSheetViewport) -> void:
 	view.param_value_edit_requested.connect(_dock._on_param_value_edit_requested)
 	view.color_swatch_edit_requested.connect(_dock._on_color_swatch_edit_requested)
 	view.param_node_drop_requested.connect(_dock._on_param_node_drop_requested)
+	view.property_dropped.connect(_dock._apply_property_drop)
 	view.variable_edit_requested.connect(_dock._on_viewport_variable_edit_requested)
 	view.comment_edit_requested.connect(_dock._open_comment_dialog)
 	view.group_edit_requested.connect(_dock._on_group_edit_requested)